refactor(steerer): extract PlanReviser (bucket 2 of 3)
Wave C bucket 2: extract the plan-install pipeline and refine-attempt
observability machinery into ``goldfive.plan_reviser.PlanReviser``.
Composition pattern with a router back-reference (``self._steerer``) —
no inheritance, no mixins. ``DefaultSteerer`` keeps the historical
public surface via thin one-liner shims so existing call sites and
tests poking the bare-attribute names continue to work unchanged.

Methods moved (22 total):

* Install entry points: ``install_initial_plan``,
  ``install_revision_for_drift``, ``install_revision_for_user_steer``,
  ``install_user_steer``, ``apply_user_steer_with_plan`` (deprecated
  back-compat shim), the shared ``_install_with_drift`` pipeline,
  and ``_build_minimal_steer_evolution``.
* The observation-only gate: ``_apply_revision`` (#254 / #255 / #258)
  — the three carve-outs (bootstrap / user-authored / discovery) are
  preserved bit-for-bit, including the ``last_addressed_revision_by_drift_key``
  watermark stamping.
* The 351-line ``_emit_plan_revised`` (#267) — copied verbatim with
  its three ``effective_dry_run`` mutation gates (supersedes-integration
  ``set_session_plan``, pending-corrections GC + queue, current_task_id
  repin) and the per-session plan-lock atomicity (#a4).
* DAG rewiring: ``_integrate_correction_supersedes`` (#251) and
  ``_repin_current_task_on_supersedes`` (#237).
* I4 fold + no-op rejection: ``_fold_runtime_terminal_statuses``,
  ``_plans_structurally_identical``.
* Refine-attempt observability: ``_emit_refine_attempted``,
  ``_emit_refine_failed``, ``observe_refine`` (the Phase-3.5-audited
  async context manager), ``_emit_plan_revised_correlation``,
  ``_new_attempt_id``.
* Per-session plan-state mutation lock: ``_get_plan_lock``,
  ``_wait_plan_stable``.
* Refine-context summaries: ``_build_refine_input_summary``,
  ``_build_refine_output_summary``.

The router keeps ownership of the shared mutable state and primitives
(``_sinks``, ``_planner``, ``_adapter``, ``_active_session_var``,
``_plan_locks``, ``_should_inject()``, ``_resolve_authored_by()``,
``_drift_annotation_id()``, ``_drift_kind_pb_value()``,
``_drift_severity_pb_value()``, ``_new_envelope()``, ``_emit()``,
``_emit_drift_detected``, ``_emit_plan_revision_transitions``, and
the not-yet-extracted drift observer methods). Cross-component calls
go through ``self._steerer.X``.

Contract preservation (referenced for the team-lead review):

* The #267 three-site dry_run gating inside ``_emit_plan_revised`` is
  copied verbatim with the original logging messages and ordering;
  ``effective_dry_run`` is resolved exactly once at the top of the
  method per the comment block that flags this fragility.
* The #254 / #255 / #258 carve-out logic in ``_apply_revision``
  (bootstrap / user-authored / discovery) is unchanged.
* The #199 ``trigger_event_id`` resolution on PlanRevised stays
  exactly as it was (pre-stamp → source annotation_id → drift.id).
* The #247 frozen-Plan ``bump_revision`` + ``set_session_plan`` shape
  is unchanged.
* The Phase 3.5 (#271) ``cancellation_stash_audited`` block in
  ``observe_refine`` is preserved verbatim including the
  ``mark_stash_completed`` call in the ``except BaseException`` arm.
* The audit catalog at ``goldfive/_state_audit.py`` already keys on
  ``(filename_suffix, qualname_suffix)`` and has no ``steerer.py``
  entries, so the move requires no catalog update.

Steerer.py: 7256 -> 5779 LOC (-1477). New ``plan_reviser.py``: 1892
LOC. One test (``test_apply_revision_silently_folds_terminal_regression``)
needed its ``caplog.at_level(..., logger="goldfive.steerer")`` widened
to ``logger="goldfive"`` so it captures the fold log line now emitted
under ``goldfive.plan_reviser``; the log message itself is unchanged.

Tests: 2089 passed, 126 skipped (zero failures); ruff clean.

Audit issues #402, #403, #404 are untouched — this refactor preserves
the existing gate behaviour for the follow-up fix PRs to land against.
